move spdf functions in analysis module
