+++ openmnglab/functions/__init__.py
@@ -1,6 +1,6 @@
 from openmnglab.functions.input.readers.dapsys_reader import DapsysReader
 from openmnglab.functions.plot.waveforms import WaveformPlot, WaveformPlotMode
 from openmnglab.functions.processing.windows import Windows
-from openmnglab.functions.processing.spdf_components import SPDFComponents
-from openmnglab.functions.processing.spdf_features import SPDFFeatures
+from openmnglab.functions.analysis.spdf_components import SPDFComponents
+from openmnglab.functions.analysis.spdf_features import SPDFFeatures
 from openmnglab.functions.processing.static_intervals import StaticIntervals
