+++ openmnglab/functions/processing/__init__.py
@@ -1,4 +1,2 @@
 from openmnglab.functions.processing.windows import Windows
-from openmnglab.functions.processing.spdf_components import SPDFComponents
-from openmnglab.functions.processing.spdf_features import SPDFFeatures
 from openmnglab.functions.processing.static_intervals import StaticIntervals
